Bug fixes & updated CMakeLists

+++ utils/hex_utils.cpp
@@ -13,7 +13,11 @@ std::vector<uint8_t> parse_hex_to_vec(const std::string& hex_str, size_t N) {
     std::vector<uint8_t> result;
     result.reserve(N);
     for (size_t i = 2; i < N * 2; i += 2) {
-        result.push_back(std::stoul(hex_str.substr(i, 2), nullptr, 16));
+        result.push_back(
+            static_cast<uint8_t>(
+                std::stoul(hex_str.substr(i, 2), nullptr, 16)
+            )
+        );
     }
     return result;
 }
